load shared view and pass data as prop in local view
