--- projects/web-app/src/store/index.js
@@ -1,5 +1,0 @@
-import { createStore } from 'vuex';
-
-export default createStore({
-  modules: {},
-});
